Use C++14 and mark time series as deprecated
The `ThreadsafeTimeseries` is deprecated and should not be used anymore
(there is a new implementation in a separate package called
"time_series").  Mark the class as deprecated so there will be compile
time warnings when it is used.

As the `[[deprecated]]` attribute was added in C++14, increase the used
standard from 11 to 14.